requirements.txt
addict==2.4.0
albumentations==1.3.1
fire==0.5.0
imgaug==0.4.0
numpy==1.23.5
opencv-python==4.9.0.80
Pillow==9.3.0
PyYAML==6.0
scikit-image==0.21.0
segmentation_models_pytorch @ git+https://github.com/qubvel/segmentation_models.pytorch@6db76a1106426ac5b55f39fba68168f3bccae7f8
torch==2.0.0.dev20230208
torchaudio==2.0.0.dev20230208
torchvision==0.15.0.dev20230208
tqdm==4.65.0